Set consistent page section leading
Before this change /learn didn't have the same line-height as the other
main sections on the site and the events section on the about-page
lacked in readability.

This sets a more consistent line-height and therefore increased
readability throughout the site, although not the yet perfect (I'd prefer
to have a proper default on the body element but as of right now that'll
mess up a lot of whitespace).